Ready version

+++ BankServer/Controllers/Types/CurrencyInfo.cs
@@ -4,5 +4,5 @@ public class CurrencyInfo
 {
     public Guid Id { get; set; }
     public string Name { get; set; }
-    public decimal Power { get; set; }
+    public long Power { get; set; }
 }
